+++ app/scripts/SeriesListItems.js
@@ -89,7 +89,9 @@ export const getSeriesItems = (
       >
         {imgTag}
         <span styleName="context-menu-span">
-          {x.name && x.name.length ? x.name : x.uid}
+          {x.options && x.options.name && x.options.name.length
+            ? x.options.name
+            : x.uid}
           {onItemMouseEnter && onItemMouseLeave ? (
             <svg styleName="play-icon">
               <use xlinkHref="#play" />