+++ app/scripts/configs/available-for-plugins.js
@@ -3,8 +3,6 @@
  */
 
 // Libraries
-import * as PIXI from 'pixi.js';
-
 import * as d3Array from 'd3-array';
 import * as d3Axis from 'd3-axis';
 import * as d3Brush from 'd3-brush';
@@ -22,6 +20,12 @@ import * as d3Zoom from 'd3-zoom';
 
 import slugid from 'slugid';
 import { mix } from '../mixwith';
+
+// Configs
+import * as configs from '.';
+
+// Utils
+import * as utils from '../utils';
 
 // Tracks
 import Annotations2dTrack from '../Annotations2dTrack';
@@ -67,12 +71,6 @@ import LruCache from '../factories';
 // Services
 import * as services from '../services';
 
-// Utils
-import * as utils from '../utils';
-
-// Configs
-import * as configs from '.';
-
 const libraries = {
   d3Array,
   d3Axis,
@@ -88,7 +86,7 @@ const libraries = {
   d3Selection,
   d3Transition,
   d3Zoom,
-  PIXI,
+  PIXI: configs.GLOBALS.PIXI,
   mix,
   slugid
 };